requirements.txt
# Core dependencies
Flask>=2.3.0
Werkzeug>=2.3.0
Jinja2>=3.1.0
MarkupSafe>=2.1.0
itsdangerous>=2.1.0
click>=8.1.0
blinker>=1.6.0

# Environment variables
python-dotenv>=1.0.0

# Face recognition (optional - can skip if using DEMO_MODE)
face-recognition>=1.3.0
dlib>=19.24.0

# DeepFace for better face recognition (from Cong-Nghe-Xu-Ly-Anh system)
deepface>=0.0.79

# YOLOv8 for face detection (optional but recommended)
ultralytics>=8.0.0

# Computer vision
opencv-python>=4.8.0
numpy>=1.24.0
pillow>=10.0.0

# Data processing
pandas>=2.0.0
python-dateutil>=2.8.0
pytz>=2023.3

# Export functionality  
openpyxl>=3.1.0
reportlab>=4.0.0

# Utilities
colorama>=0.4.6

# Internationalization (i18n)
Flask-Babel>=4.0.0
